[Fix] migrate create_bookmarks.rb

+++ db/migrate/20210908034109_create_bookmarks.rb
@@ -1,8 +1,8 @@
 class CreateBookmarks < ActiveRecord::Migration[5.2]
   def change
     create_table :bookmarks do |t|
-      t.references :user, type: :bigint, foreign_key: true, null: false
-      t.references :manual, type: :bigint, foreign_key: true, null: false
+      t.references :user, type: :integer, foreign_key: true, null: false
+      t.references :manual, type: :integer, foreign_key: true, null: false
 
       t.timestamps
     end
